Move SFENConverter to /board_notation.
